requirements.txt
runpod
boto3
Pillow
